fix: Load Plant Sprites On Load
Previously, a soft pointer was used to allow the assets to be loaded when needed. It is more ideal that the assets are available on demand without needing to be loaded in real time.

For that reason, the pointer is being reverted to a regular object pointer. Soft pointer can be considered when there are too many plant types, which there aren't many.